Ica Fields, reduced, with NDVI
This CSV contains NDVI values for 2019-2021 related to each of the segmented fields in Ica. The no data values were filled in with linear interpolation. Null values at the beginning or end of the dataset took the closest non-null value. 
The segmented ICA fields can be related back to the spatial dataset with the uniqueID column called "label".